requirements.txt
langchain
langchain-groq
langchain-community
chromadb
pypdf
python-dotenv
sentence-transformers